gemini_examples/requirements.txt
google-genai
gTTS
SpeechRecognition
pyaudio
opencv-python
sox
pydub
python-dotenv
openwakeword
onnxruntime
numpy
